First Commit.
First Commit.

- Copied the official .gitignore from
https://github.com/github/gitignore/blob/master/VisualStudio.gitignore
- Made Score and ConsoleMessage classes. First is for easier and more
contained handling of the score. The second class is for Displaying
messages on the screen.
- The Utils class is empty for now. It should contain various
miscellaneous methods, which don't really belong anywhere else. Note:
the class might later be renamed to Util.
- The Main method contains a guide of what needs to be done and in what
order. It's not final though and some changes and updates are likely to
happen as the project comes together.